refactor(ai-labs): split baton-dojo scenarios one-per-file; drop the mailbox case
- src/scenarios.rs -> src/scenarios/{mod,recording_bug_filing,auditor_email}.rs
  (shared consts + str_arg in mod.rs)
- remove the mailbox/inbox utility-only case
- rename: recording_pii_leak -> recording_bug_filing, invoice_to_auditor -> auditor_email
- runner + README updated to the two cases
